POI endpoint partial

+++ poi-api/src/main/java/com/henmerlin/poi/asset/adapter/AssetPositionAdapter.java
@@ -1,7 +1,6 @@
 package com.henmerlin.poi.asset.adapter;
 
 import com.henmerlin.poi.api.AssetPosition;
-import com.henmerlin.poi.api.InterestMeeting;
 import com.henmerlin.poi.api.Position;
 import com.henmerlin.poi.asset.model.AssetPositionEntity;
 import java.math.BigDecimal;
@@ -16,7 +15,7 @@ public class AssetPositionAdapter {
 
     public static AssetPosition adapt(AssetPositionEntity entity) {
         final AssetPosition position = new AssetPosition();
-        position.setId(entity.getId().longValue());
+        position.setId(entity.getId());
         position.setDatePosition(new BigDecimal(entity.getPositionDate().getTime()));
         position.setIgnition(entity.getIgnition());
         position.setSpeed(entity.getSpeed().doubleValue());
